fixed tests 5
